+++ cache/config/config.exs
@@ -14,7 +14,13 @@ config :cache, Cache.PromEx,
   drop_metrics_groups: [],
   grafana: :disabled
 
-config :cache, Cache.Repo, busy_timeout: 10_000
+config :cache, Cache.Repo,
+  busy_timeout: 30_000,
+  journal_mode: :wal,
+  synchronous: :normal,
+  temp_store: :memory,
+  queue_target: 1_000,
+  queue_interval: 1_000
 
 config :cache, CacheWeb.Endpoint,
   url: [host: "localhost"],